refactor(dof): move check_dof_daily into dof_ingest to clear file-size gate
After merging main (which added the CONAMER ingest task to tasks.py in
#140), tasks.py hit 816 lines (>800 gate). Move the whole check_dof_daily
task alongside its materialization helpers in dof_ingest.py; tasks.py
re-exports it so Celery discovery and the Beat schedule (by task name) are
unchanged. tasks.py now 744 lines.
