+++ apps/scraper/scheduling/dof_ingest.py
@@ -9,6 +9,8 @@ direct PDF, so an operator validates materialization before enabling in prod.
 
 import logging
 
+from celery import shared_task
+
 logger = logging.getLogger(__name__)
 
 
@@ -67,3 +69,87 @@ def _materialize_dof_changes(changes):
             failed += 1
             logger.exception("DOF auto-ingest failed for %s", metadata.get("id"))
     return materialized, failed
+
+
+@shared_task(name="dataops.check_dof_daily")
+def check_dof_daily():
+    """Check today's DOF edition for law changes.
+
+    Runs daily at 7 AM via Celery Beat. Fetches the DOF index, detects
+    reforms/new laws/abrogations, and — when DOF_AUTO_INGEST_ENABLED is set —
+    materializes new-law/reform publications through the ingestion pipeline.
+    With the flag off it detects + logs only.
+    """
+    import datetime
+
+    # Pass existing law names so the scraper can match against actual DB laws
+    from apps.api.models import Law
+    from apps.scraper.federal.dof_daily import DofScraper
+
+    existing_laws = list(Law.objects.values_list("name", flat=True))
+
+    scraper = DofScraper(date=datetime.date.today())
+    results = scraper.run(existing_laws=existing_laws)
+
+    entries = results.get("entries", [])
+    changes = results.get("changes", [])
+
+    materialized, materialize_failed = _materialize_dof_changes(changes)
+
+    # Log to AcquisitionLog
+    try:
+        from apps.scraper.dataops.models import AcquisitionLog
+
+        log_entry = AcquisitionLog.objects.create(
+            operation="dof_daily_check",
+            parameters={
+                "date": str(datetime.date.today()),
+                "existing_laws_count": len(existing_laws),
+                "detected": len(changes),
+                "changes": [
+                    {
+                        "change_type": c.get("change_type"),
+                        "title": c.get("title", "")[:200],
+                    }
+                    for c in changes[:20]
+                ],
+            },
+            found=len(entries),
+            downloaded=materialized,
+            failed=materialize_failed,
+            # `ingested` is now an accurate count of laws written to the DB
+            # (0 when auto-ingest is off), not the prior detected-count misnomer.
+            ingested=materialized,
+        )
+        if changes:
+            log_entry.error_summary = (
+                f"{len(changes)} law changes detected, "
+                f"{materialized} materialized, {materialize_failed} failed"
+            )
+        log_entry.finish()
+    except Exception:
+        # Log persistence failure — don't fail the actual DOF check, but
+        # surface it so AcquisitionLog gaps don't go unnoticed.
+        logger.exception("Failed to persist AcquisitionLog for dof_daily_check")
+
+    if changes:
+        logger.warning(
+            "DOF daily: %d entries, %d law changes detected",
+            len(entries),
+            len(changes),
+        )
+        for change in changes[:10]:
+            logger.warning(
+                "  [%s] %s", change.get("change_type", "?"), change.get("title", "?")
+            )
+    else:
+        logger.info("DOF daily: %d entries, no law changes detected", len(entries))
+
+    return {
+        "date": str(datetime.date.today()),
+        "total_entries": len(entries),
+        "law_changes": len(changes),
+        "materialized": materialized,
+        "materialize_failed": materialize_failed,
+        "changes": changes[:20],
+    }
